move signingSession to tree folder
