+++ src/identity/index.ts
@@ -1,5 +1,5 @@
 import { Transaction } from "@scure/btc-signer";
-import { SignerSession } from "../signingSession";
+import { SignerSession } from "../tree/signingSession";
 
 export interface Identity {
     // if inputIndexes is not provided, try to sign all inputs
